+++ src/main/java/Duke/Exception/EmptyArgumentException.java
@@ -1,2 +1,4 @@
+package Duke.Exception;
+
 public class EmptyArgumentException extends Exception {
 }
