+++ src/main/java/Duke/Exception/WrongArgumentException.java
@@ -1,2 +1,4 @@
+package Duke.Exception;
+
 public class WrongArgumentException extends Exception {
 }
