+++ src/main/java/Duke/Tasks/Deadline.java
@@ -1,3 +1,5 @@
+package Duke.Tasks;
+
 public class Deadline extends Task {
     private String time;
