+++ src/main/java/Duke/Tasks/Event.java
@@ -1,3 +1,5 @@
+package Duke.Tasks;
+
 public class Event extends Task {
     private String time;
