+++ src/main/java/Duke/Tasks/Todo.java
@@ -1,3 +1,5 @@
+package Duke.Tasks;
+
 public class Todo extends Task {
     public Todo (String description) {
         super(description);
