--- src/main/java/TaskType.java
@@ -1,3 +1,0 @@
-public enum TaskType {
-    TODO, DEADLINE, EVENT, OTHER
-}
