+++ text-ui-test/EXPECTED.TXT
@@ -4,7 +4,7 @@ ____________________________________________________________
 | | | | | | | |/ / _ \
 | |_| | |_| |   <  __/
 |____/ \__,_|_|\_\___|
-Hello! I'm Duke
+Hello! I'm Duke.Duke
 ____________________________________________________________
 
 What can I do for you?
